+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 20) Fix blank uuid in console log
 21) Add Polish translation, thanks to @LimakXRobczuk
 22) Fix [#51](https://github.com/NikitaCartes/EasyAuth/issues/51) that sometimes online player treated as offline
+23) Added options to allow certain commands to be used before login
 
 ### 2.2.2, Minecraft 1.17, 1.17.1, 1.18[.1], 1.18.2